Attacking the encyclopedia grid rewrite
