+++ src/assets/data/levels/classic/level4.json
@@ -14,6 +14,7 @@
         },
         "element": "Laser",
         "rotation": 0,
+        "polarization": 90,
         "frozen": true,
         "active": true,
         "energized": false
