+++ src/components/EncyclopediaPage/EncyclopediaBoard.vue
@@ -64,7 +64,13 @@
 import { Vue, Prop, Component } from 'vue-property-decorator';
 import { Grid, Cell } from '@/engine/classes';
 import Particle from '@/engine/Particle';
-import { ParticleInterface, GridInterface, PolEnum, IndicatorInterface } from '@/engine/interfaces';
+import {
+  ParticleInterface,
+  GridInterface,
+  PolEnum,
+  IndicatorInterface,
+  DirEnum
+} from '@/engine/interfaces';
 import AppPhoton from '@/components/AppPhoton.vue';
 import BoardDots from '@/components/Board/BoardDots.vue';
 import BoardLasers from '@/components/Board/BoardLasers.vue';
@@ -89,10 +95,11 @@ export default class EncyclopediaBoard extends Vue {
   @Prop({ default: () => dummyGridInterface }) gridObj!: GridInterface;
   @Prop({ default: () => 10 }) readonly maxSteps!: number;
   @Prop({ default: () => 2 }) readonly defaultStep!: number;
-  @Prop({ default: () => [] }) readonly indicators!: IndicatorInterface[];
+  @Prop({ default: () => [{ x: 0, y: 1, dirStr: DirEnum['>'], polStr: PolEnum.H }] })
+  readonly indicators!: IndicatorInterface[];
   @Prop({ default: false }) readonly exactSteps!: boolean;
+  @Prop({ default: 64 }) readonly tileSize!: number;
 
-  tileSize = 64; // sounds like a prop or constant
   grid = new Grid(
     this.gridObj.rows,
     this.gridObj.cols,
@@ -118,11 +125,13 @@ export default class EncyclopediaBoard extends Vue {
     } else {
       throw new Error('EncyclopediaBoard not yet prepared for more photons.');
     }
+
     if (this.exactSteps) {
       this.simulation.nextFrames(this.maxSteps, -1);
     } else {
       this.simulation.nextFrames(this.maxSteps);
     }
+
     this.selectedFrameId = Math.min(this.selectedFrameId, this.simulation.frames.length - 1);
   }
