+++ src/components/EncyclopediaPage/EncyclopediaMatrix.vue
@@ -41,51 +41,7 @@
 
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-
-const TAU = 2 * Math.PI;
-
-/**
- * Stolen from https://stackoverflow.com/questions/36721830/convert-hsl-to-rgb-and-hex
- * Alternatively: d3.hsl
- */
-function hslToHex(hParam: number, sParam: number, lParam: number) {
-  let h = hParam;
-  let s = sParam;
-  let l = lParam;
-  h /= 360;
-  s /= 100;
-  l /= 100;
-  let r;
-  let g;
-  let b;
-  if (s === 0) {
-    r = l;
-    g = l;
-    b = l; // achromatic
-  } else {
-    const hue2rgb = (pParam: number, qParam: number, tParam: number) => {
-      const p = pParam;
-      const q = qParam;
-      let t = tParam;
-      if (t < 0) t += 1;
-      if (t > 1) t -= 1;
-      if (t < 1 / 6) return p + (q - p) * 6 * t;
-      if (t < 1 / 2) return q;
-      if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
-      return p;
-    };
-    const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-    const p = 2 * l - q;
-    r = hue2rgb(p, q, h + 1 / 3);
-    g = hue2rgb(p, q, h);
-    b = hue2rgb(p, q, h - 1 / 3);
-  }
-  const toHex = (x: number) => {
-    const hex = Math.round(x * 255).toString(16);
-    return hex.length === 1 ? `0${hex}` : hex;
-  };
-  return `#${toHex(r)}${toHex(g)}${toHex(b)}`;
-}
+import { hslToHex, colorComplex } from '@/engine/Helpers';
 
 @Component
 export default class EncyclopediaOperatorViewer extends Vue {
@@ -93,24 +49,18 @@ export default class EncyclopediaOperatorViewer extends Vue {
   @Prop({ default: () => 600 }) private height!: number;
   @Prop({ default: () => 40 }) private size!: number;
   @Prop({ default: () => 40 }) private margin!: number;
-  // @Prop() private height = 600  // generates error
+  @Prop({ default: () => [] }) private labelsIn!: number[];
+  @Prop({ default: () => [] }) private labelsOut!: number[];
   @Prop({ default: () => [] }) private matrixElements!: {
     i: number;
     j: number;
     re: number;
     im: number;
   }[];
-  @Prop({ default: () => [] }) private labelsIn!: number[];
-  @Prop({ default: () => [] }) private labelsOut!: number[];
 
-  scale = (i: number) => i * this.size;
-
-  // https://github.com/stared/quantum-game/blob/master/js/transition_heatmap.js
-  colorComplex = (re: number, im: number) => {
-    const angleInDegrees = ((Math.atan2(im, re) * 360) / TAU + 360) % 360;
-    const r = Math.sqrt(re * re + im * im); // for pure color it should be always 1
-    return hslToHex(angleInDegrees, 100, 100 - 50 * r);
-  };
+  scale(i: number): number {
+    return i * this.size;
+  }
 
   tileMouseOver(d: { i: number; j: number; re: number; im: number }) {
     this.$emit('columnMouseover', d.i);
@@ -118,7 +68,6 @@ export default class EncyclopediaOperatorViewer extends Vue {
 }
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped lang="scss">
 .operator-viewer {
   display: inline-block;
